+++ checkpoint/orbax/checkpoint/_src/metadata/step_metadata_serialization.py
@@ -18,6 +18,7 @@ import dataclasses
 from typing import Any
 
 from absl import logging
+from etils import epath
 from orbax.checkpoint._src.logging import step_statistics
 from orbax.checkpoint._src.metadata import checkpoint
 from orbax.checkpoint._src.metadata import metadata_serialization_utils as utils
@@ -31,6 +32,17 @@ CheckpointHandlerTypeStr = checkpoint.CheckpointHandlerTypeStr
 CompositeItemMetadata = checkpoint.CompositeItemMetadata
 SingleItemMetadata = checkpoint.SingleItemMetadata
 StepStatistics = step_statistics.SaveStepStatistics
+
+
+def get_step_metadata(path: epath.PathLike) -> StepMetadata:
+  """Returns StepMetadata for a given checkpoint directory."""
+  metadata_file_path = checkpoint.step_metadata_file_path(path)
+  serialized_metadata = checkpoint.metadata_store(enable_write=False).read(
+      metadata_file_path
+  )
+  if serialized_metadata is None:
+    raise ValueError(f'Step metadata not found at {metadata_file_path}')
+  return deserialize(serialized_metadata)
 
 
 def serialize(metadata: StepMetadata) -> SerializedMetadata: